Clarify compile-time errors
An example like

    let foo = "hello!";
    foo();

would gets newbie readers very confused: if all you do is "run the
code", then doing so with plain JS or with TS gets you an error anyway,
so what's the big fuss?  Clarify this in the text, and also add a
`function` so that plain JS wouldn't complain, to clarify it further.